Some more Refactoring
Some more minor refactoring is done with the code. This refactoring is
done without changing its output for the code. There were many changes
done to the code so that it does not look cluttered and makes it even
more clean and convenient to understand.
All of it's points are listed below:-
* We replaced all the cities_data variable used afterwards with the variable which was assigned to it
* We replaced all the summaries used afterwards with all those variables which were assigned them
* Also changed the method in which two coloumns each were deleted each from the tables ny and chi in our code
* Appropriately changed and arranged some comments in different parts of the code
* Reformated the whole code once again in a better manner